+++ activesupport/lib/active_support/cache/serializer_with_fallback.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require "zlib"
 require "active_support/core_ext/kernel/reporting"
 
 module ActiveSupport
